Move zeroPadding.js
